requirements.txt
openai>=1.40.0
chromadb>=0.5.5
python-dotenv>=1.0.1
rich>=13.7.1
tiktoken>=0.7.0
streamlit>=1.37.0
pydantic>=2.7.0